Clean up of code to match rubocop rules.
Not sure how terse to go with some of the code. I'm still not 100% with
accepted Ruby idioms.
