+++ gem_city.rb
@@ -1,0 +1,44 @@
+# This class represents the town of GemCity
+# This is a town riddled with crime but we can find out how happy the town is
+class GemCity
+  attr_reader :population
+  def initialize
+    @people = {
+      thieves:  5,
+      Officers: 1
+    }
+    @population = 50
+  end
+
+  def thieves(thieves_number = @people[:thieves])
+    @people[:thieves] = thieves_number
+  end
+
+  def officers
+    @people[:Officers]
+  end
+
+  def officers=(officers)
+    @people[:Officers] = officers
+  end
+
+  def happiness_of_town
+    # happiness is random... people don't know what they want!
+    happiness = 0
+    (1..population).each do
+      happiness += rand((100 - successful_crime_rate)..100)
+    end
+    happiness / 100
+  end
+
+  def successful_crime_rate
+    thieves = @people[:thieves]
+    officers = @people[:Officers]
+    odds_percent = 0
+    if thieves > 0 && officers <= thieves
+      odds = 1 - officers.to_f / thieves.to_f
+      odds_percent = odds * 100
+    end
+    odds_percent
+  end
+end
